# Improved Performance of loadClassesByScalaName