Refactored SourceLookUpProvider.loadClassesByScalaName() to be more efficient. 

The issue presented itself in 4.0.0 (related to the HotCodeReplace feature). The debug loading times became really slow.

On my large project, the loading went from 75s back to a normal 12s. (12s is about what I saw in v3.1.6).  

This should resolve https://github.com/scalameta/metals/issues/6544