@adpi2 2024-07-08
Thanks for your contribution @crt-31 

Using the `SourceLookUpProviderStats` we can clearly see the speedup:

|                | Before  | After   |
|----------------|---------|---------|
| ubuntu-latest  | 1510 ms | 830 ms  |
| windows-latest | 2299 ms | 1297 ms |
| macOS-latest   | 1682 ms | 633 ms  |

I also wonder if we could make it even faster using a `mutable.Map`.